Move Buy Tickets button from Museum Info to Info
