+++ info/src/main/res/layout/fragment_museum_information.xml
@@ -101,25 +101,13 @@
                     android:layout_height="wrap_content"
                     android:layout_marginStart="@dimen/marginDouble"
                     android:layout_marginTop="@dimen/marginDouble"
+                    android:layout_marginBottom="@dimen/marginQuad"
                     android:fontFeatureSettings="tnum"
+                    app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toBottomOf="@id/museumAddress"
                     tools:text="222-346-1234" />
-
-                <Button
-                    android:id="@+id/buyTickets"
-                    style="@style/DetailButton"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/marginDouble"
-                    android:layout_marginTop="@dimen/marginQuad"
-                    android:layout_marginEnd="@dimen/marginDouble"
-                    android:background="@color/brownishOrange"
-                    android:text="@string/buyTickets"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/museumPhone" />
 
             </android.support.constraint.ConstraintLayout>
         </android.support.v4.widget.NestedScrollView>
